% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassificationHistogram.R
\name{ClassificationHistogram}
\alias{ClassificationHistogram}
\title{ClassificationHistogram.}
\usage{
ClassificationHistogram(out_classific = NULL, breaks = seq(0, 1, 0.05), ...)
}
\arguments{
\item{out_classific}{Output of \link{ClassificationWrapper}.}

\item{breaks}{Breaks for histogram.}

\item{...}{Passed on to \code{par}. Useful to adjust \code{cex}.}
}
\value{
Returns NULL invisibly.
}
\description{
\code{ClassificationHistogram} will plot the results of \link{ClassificationWrapper}.
}
\details{
No further details.
}
