% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolarPlot.R
\name{PolarCoordHeterPlot}
\alias{PolarCoordHeterPlot}
\title{PolarCoordHeterPlot.}
\usage{
PolarCoordHeterPlot(
  foo = foo,
  gt = c("B73", "B73xMo17", "Mo17"),
  use_exp = TRUE,
  exp_fac = 0.45,
  thr = 1,
  plot_lab = c("none", "text", "graph"),
  col = NULL
)
}
\arguments{
\item{foo}{Data matrix with measurement values (traits in rows and genotypes in columns).}

\item{gt}{Character vector of length=3 indicating P1, F and P2 to be filtered by column name from foo.}

\item{use_exp}{.}

\item{exp_fac}{Expansion factor to increase figure size.}

\item{thr}{Alpha level used in ANOVA to filter insignificant rows. Keep thr=1 to include all matrix rows.}

\item{plot_lab}{Show 'text' or 'graph' style labels of the polar sections (or kepp 'none' to omit).}

\item{col}{Provide a color vector of length nrow(foo).}
}
\value{
Will generate a plot in polar coordinates and return the x/y coordinates of the data points invisibly.
}
\description{
\code{PolarCoordHeterPlot} will draw a plot in polar coordinates
   visualizing heterosis effects according to a layout by Swanson-Wagner,
   where plot radius represents log2 of fold change between lowest and highest
   genotype and plot angle represents the ratio between lowest, intermediate
   and highest genotype.
}
\details{
See examples.
}
\examples{
gt <- c("P1","P1xP2","P2")
set.seed(0)
foo <- matrix(rnorm(150), nrow = 10, dimnames = list(paste0("M",1:10), sample(rep(gt, 5))))
foo[1:4,1:6]
PolarCoordHeterPlot(foo=foo, gt=gt, exp_fac=1)
PolarCoordHeterPlot(foo=foo, gt=gt, plot_lab="text")
PolarCoordHeterPlot(foo=foo, gt=c("P2","P1xP2","P1"), plot_lab="text")
PolarCoordHeterPlot(foo=foo, gt=c("P2","P1xP2","P1"), plot_lab="graph")
PolarCoordHeterPlot(foo=foo, gt=gt, exp_fac=1, col=1:10)
PolarCoordHeterPlot(foo=foo, gt=gt, exp_fac=1, col=1:10, thr=0.5)
PolarCoordHeterPlot(foo=foo, gt=gt, plot_lab="graph", exp_fac=1, col=1:10, thr=0.5)
}
