% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolomicsBasics-data.R
\docType{data}
\name{raw}
\alias{raw}
\title{Metabolomics data set}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 120 rows and 112 columns.
}
\source{
<https://onlinelibrary.wiley.com/doi/10.1111/j.1365-313X.2011.04689.x>
}
\usage{
raw
}
\description{
This data set contains a matrix of log10-transformed ion intensities from a maize root metabolomics study for in total 112 metabolites in 120 samples.
}
\author{
Jan Lisec \email{jan.lisec@bam.de}
}
\keyword{data}
