% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msdial.R
\name{read_msdial}
\alias{read_msdial}
\title{read_msdial}
\usage{
read_msdial(file, sam = NULL)
}
\arguments{
\item{file}{The alignment file as exported from MSDial.}

\item{sam}{An additional sample list (data.frame) where one column can be matched to the file names provided by MSDial.}
}
\value{
A list with at least 3 components (`sam`, `met` and `raw`) containing the relevant information on the experiment.
}
\description{
read_msdial
}
