% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalib_mzML.R
\name{recalib_mzML}
\alias{recalib_mzML}
\title{recalib_mzML}
\usage{
recalib_mzML(
  infile,
  outfile = sub("\\\\.[mM][zZ][mM][lL]$", "_calib.mzML", infile, ignore.case = TRUE),
  shift_da = -0.019,
  apply_to = c("all", "MS1", "MS2")
)
}
\arguments{
\item{infile}{mzML file path.}

\item{outfile}{If no target file is specified, input file name will be used (extended by '_calib').}

\item{shift_da}{The desired m/z shift in Dalton.}

\item{apply_to}{Use keyword 'all' to modify MS1 and MS2 or specify selectively.}
}
\value{
The modified XML invisibly (for potential checks).
}
\description{
If calibration fails, mass drift may occur within a measurement
    batch, leading to m/z values in the raw data being offset with respect to
    the true value. This often can be fixed using software tools provided by
    the vendor. However, it may become necessary to re-calibrate mzML data
    after conversion from (incorrect) raw data. This function allows to shift
    m/z values in MS1 and/or MS2 scans of mzML files by a fixed value.
    Except for base_peak and precursor values assigned to spectra, which are
    updated, all other nodes of the mzML file will be preserved. The binary
    encoding (32-bit or 64-bit of the measurement data is respected by the
    function.)
}
\examples{
\dontrun{
  fld <- "C:/Users/jlisec/Documents/Projects/PFAS_nontargeted/TSo_GMW/mzML/original/"
  fls <- dir(fld, pattern = "mzML$", full.names = TRUE)
  fl <- fls[1]
  for (fl in fls) {
    print(fl)
    recalib_mzML(infile = fl, shift_da = -0.012)
  }
}
}
