% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam-data.R
\docType{data}
\name{sam}
\alias{sam}
\title{Sample table}
\description{
This data frame contains the sample definition of 120 samples according to the rows of \link{raw}.
}
\references{
\url{http://dx.doi.org/10.1111/j.1365-313X.2011.04689.x}
}
\author{
Jan Lisec \email{jan.lisec@charite.de}
}
\keyword{data}
