% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_format_converter.R
\name{spectra_format_converter}
\alias{spectra_format_converter}
\title{spectra_format_converter.}
\usage{
spectra_format_converter(txt = NULL, m_prec = 3, i_prec = 0)
}
\arguments{
\item{txt}{Sample table.}

\item{m_prec}{Mass precision of output spectrum.}

\item{i_prec}{Intensity precision of output spectrum.}
}
\value{
Matrix with mz and int columns.
}
\description{
\code{spectra_format_converter} will generate a matrix with mz and int
  columns out of a text representation of a spectrum.
}
\details{
See examples.
}
\examples{
spectra_format_converter(txt = "57.1:100 58.0001:10")
spectra_format_converter(txt = "58.0001:10 57.1:100", m_prec = 4)
}
