% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_labels.R
\name{unique_labels}
\alias{unique_labels}
\title{unique_labels.}
\usage{
unique_labels(sam = NULL, g = NULL)
}
\arguments{
\item{sam}{Sample table.}

\item{g}{Either column name from sam containing factor column or factor of same length as sam.}
}
\value{
Dataframe with group levels names and their color and plotting character specification.
}
\description{
\code{unique_labels} will generate a dataframe with color and
  plotting character specification out of a sample table definition.
}
\details{
If a color/symbol specification exists for a sample set containing
  replicate groups this function will help in retrieving this information
  per group which is useful in boxplot or legend functions (cf. examples).
}
\examples{
utils::data(raw, package = "MetabolomicsBasics")
utils::data(sam, package = "MetabolomicsBasics")
unique_labels(sam = sam, g = "GT")
}
