% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassificationHistogram.R
\name{ClassificationHistogram}
\alias{ClassificationHistogram}
\title{ClassificationWrapper.}
\usage{
ClassificationHistogram(out_classific = NULL, breaks = seq(0, 1, 0.05), ...)
}
\arguments{
\item{out_classific}{Output of \link{ClassificationWrapper}.}

\item{breaks}{Breaks for histogram.}

\item{...}{Passed on to \code{par}. Useful to adjust \code{cex}.}
}
\value{
Classification results as list.
}
\description{
\code{ClassificationWrapper} will do classification using SVM's and/or Decision Trees including cross validation.
}
\details{
not yet
}
\examples{
# check the examples in \code{\link{ClassificationWrapper}}

}
