% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolomicsBasics-data.R
\docType{data}
\name{met}
\alias{met}
\title{Metabolite table}
\format{
An object of class \code{data.frame} with 112 rows and 2 columns.
}
\source{
<https://onlinelibrary.wiley.com/doi/10.1111/j.1365-313X.2011.04689.x>
}
\usage{
met
}
\description{
This data frame contains the metabolite definition of 112 metabolites according to the cols of \link{raw}.
}
\author{
Jan Lisec \email{jan.lisec@bam.de}
}
\keyword{data}
