% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met-data.R
\docType{data}
\name{met}
\alias{met}
\title{Metabolite table}
\description{
This data frame contains the metabolite definition of 112 metabolites according to the cols of \link{raw}.
}
\references{
\url{http://dx.doi.org/10.1111/j.1365-313X.2011.04689.x}
}
\author{
Jan Lisec \email{jan.lisec@charite.de}
}
\keyword{data}
