% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-data.R
\docType{data}
\name{raw}
\alias{raw}
\title{Metabolomics data set}
\format{An object of class \code{matrix} with 120 rows and 112 columns.}
\usage{
raw
}
\description{
This data set contains log10-transformed raw data of a maize root metabolomics study for in total 112 metabolites in 120 samples.
}
\references{
\url{http://dx.doi.org/10.1111/j.1365-313X.2011.04689.x}
}
\author{
Jan Lisec \email{jan.lisec@charite.de}
}
\keyword{data}
