% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolomicsBasics-data.R
\docType{data}
\name{sam}
\alias{sam}
\title{Sample table}
\format{
An object of class \code{data.frame} with 120 rows and 10 columns.
}
\source{
<https://onlinelibrary.wiley.com/doi/10.1111/j.1365-313X.2011.04689.x>
}
\usage{
sam
}
\description{
This data frame contains the sample definition of 120 samples according to the rows of \link{raw}.
}
\author{
Jan Lisec \email{jan.lisec@bam.de}
}
\keyword{data}
